Improve logging of failed subprocesses
