+++ include/linuxdeploy/plugin/plugin.h
@@ -5,7 +5,7 @@
 #include <string>
 
 // local includes
-#include "linuxdeploy/core/log.h"
+#include "linuxdeploy/log/log.h"
 #include "linuxdeploy/plugin/exceptions.h"
 
 #pragma once
