+++ src/plugin/plugin.cpp
@@ -3,20 +3,14 @@
 #include <regex>
 #include <set>
 #include <string>
-#include <vector>
-
-// library headers
-#include <fnmatch.h>
 
 // local headers
-#include "linuxdeploy/core/log.h"
-#include "linuxdeploy/plugin/base.h"
+#include "linuxdeploy/log/log.h"
 #include "linuxdeploy/plugin/plugin.h"
 #include "linuxdeploy/util/util.h"
 #include "plugin_type0.h"
 
-using namespace linuxdeploy::core;
-using namespace linuxdeploy::core::log;
+using namespace linuxdeploy::log;
 
 namespace fs = std::filesystem;
